Added Final Changes

+++ Backend/routes/posts.js
@@ -32,7 +32,8 @@ router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   
   post.save().then((createdPost) => {
@@ -45,6 +46,11 @@ router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res
         imagePath: createdPost.imagePath
       }
     });
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Creating Post Failed"
+    })
   });
 });
 
@@ -58,15 +64,19 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
         _id: req.params.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath 
+        imagePath: imagePath,
+        creator: req.userData.userId 
     });
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-        console.log(result);
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+      if(result.n > 0) {
         res.status(200).json({message: 'Update successful!'});
+      } else {
+        res.status(401).json({message: 'Not Authorized!'});
+      }
     })
 })
 
-router.get("", (req, res, next) => {
+router.get("/", (req, res, next) => {
   const pageSize = +req.query.pageSize;
   const currentPage = +req.query.page;
   const postQuerry = Post.find();
@@ -99,9 +109,12 @@ router.get("/:id", (req, res, next) => {
 
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
-    console.log(result);
-    res.status(200).json({ message: "Post Deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => { 
+    if(result.deletedCount) {
+        res.status(200).json({message: 'Deleted successful!'});
+      } else {
+        res.status(401).json({message: 'Not Authorized!'});
+      }
   });
 });
